@@ -4,8 +4,10 @@ import './Footer.less';
 
 import GitHub from './GitHub';
 
-export default () => (
-  <footer>
-    <GitHub />
-  </footer>
-);
+export default function Footer() {
+  return (
+    <footer>
+      <GitHub />
+    </footer>
+  );
+}
