@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import mergeClassNames from 'merge-class-names';
 
@@ -11,22 +11,16 @@ import Subsection from './Subsection';
 import Method from './Method';
 import Initiator from './Initiator';
 
-export default class Section extends Component {
-  static propTypes = {
-    advanced: PropTypes.bool,
-    children: PropTypes.node,
-    col: PropTypes.number.isRequired,
-    colspan: PropTypes.number,
-    name: PropTypes.string.isRequired,
-  };
+export default function Section(props) {
+  const {
+    advanced,
+    children,
+    col,
+    colspan,
+    name,
+  } = props;
 
-  static defaultProps = {
-    colspan: 1,
-  }
-
-  renderChildren() {
-    const { children, col: sectionCol } = this.props;
-
+  function renderChildren() {
     // If we're creating a section containing subsections, we don't need to create one.
     if (!children.find(el => el.type === Method || el.type === Initiator)) {
       return React.Children.map(
@@ -34,8 +28,8 @@ export default class Section extends Component {
         child => React.cloneElement(
           child,
           Object.assign(
-            { sectionCol },
-            this.props,
+            { sectionCol: col },
+            props,
             child.props,
           ),
         ),
@@ -43,46 +37,52 @@ export default class Section extends Component {
     }
 
     return (
-      <Subsection {...this.props} />
+      <Subsection {...props} />
     );
   }
 
-  render() {
-    const {
-      advanced, col, colspan, name,
-    } = this.props;
+  const gridColumn = `${col + 1} / span ${colspan}`;
 
-    const gridColumn = `${col + 1} / span ${colspan}`;
-
-    return (
-      <>
-        <section
-          className={mergeClassNames('Section', advanced && 'Section--advanced')}
-          style={{
-            gridColumn,
-            gridRow: advanced ? '1 / span 23' : '1 / span 14',
-          }}
-        />
-        <div
-          className={mergeClassNames('Section__highlight', advanced && 'Section__highlight--advanced')}
-          style={{
-            gridColumn,
-            gridRow: advanced ? '16 / span 8' : '8 / span 7',
-          }}
-        />
-        <h3
-          className="Section__title"
-          style={{
-            gridColumn,
-            gridRow: 1,
-          }}
-        >
-          <T>
-            {name}
-          </T>
-        </h3>
-        {this.renderChildren()}
-      </>
-    );
-  }
+  return (
+    <>
+      <section
+        className={mergeClassNames('Section', advanced && 'Section--advanced')}
+        style={{
+          gridColumn,
+          gridRow: advanced ? '1 / span 23' : '1 / span 14',
+        }}
+      />
+      <div
+        className={mergeClassNames('Section__highlight', advanced && 'Section__highlight--advanced')}
+        style={{
+          gridColumn,
+          gridRow: advanced ? '16 / span 8' : '8 / span 7',
+        }}
+      />
+      <h3
+        className="Section__title"
+        style={{
+          gridColumn,
+          gridRow: 1,
+        }}
+      >
+        <T>
+          {name}
+        </T>
+      </h3>
+      {renderChildren()}
+    </>
+  );
 }
+
+Section.propTypes = {
+  advanced: PropTypes.bool,
+  children: PropTypes.node,
+  col: PropTypes.number.isRequired,
+  colspan: PropTypes.number,
+  name: PropTypes.string.isRequired,
+};
+
+Section.defaultProps = {
+  colspan: 1,
+};
