@@ -2,7 +2,8 @@
 
 import t from './t.js';
 import T from './T.jsx';
+import useTranslation from './useTranslation';
 
-export { t };
+export { t, useTranslation };
 
 export default T;
