@@ -1,33 +1,37 @@
 import '@babel/polyfill';
+
 import { defaultLocale, getMatchingLocale, languageFiles } from './i18n';
 
-export default async (string, args, locale) => {
-  if (!args) {
-    // eslint-disable-next-line no-param-reassign
-    args = {};
-  }
-
+export default async function t(string, args, locale) {
   if (!locale) {
     // eslint-disable-next-line no-param-reassign
     locale = getMatchingLocale();
   }
 
-  const getTranslatedString = async () => {
+  async function getTranslatedString() {
     if (locale !== defaultLocale) {
-      const languageFile = await languageFiles[locale];
-      if (!languageFile) {
+      let languageFile;
+      try {
+        languageFile = await languageFiles[locale];
+      } catch (error) {
         // eslint-disable-next-line no-console
         console.error(`Unable to load locale: ${locale}`);
         return string;
       }
+
       if (typeof languageFile[string] === 'string') {
         return languageFile[string];
       }
     }
+
     return string;
-  };
+  }
 
   const rawString = await getTranslatedString();
+
+  if (!args) {
+    return rawString;
+  }
 
   let finalString = rawString;
   Object.entries(args).forEach(([key, value]) => {
@@ -35,4 +39,4 @@ export default async (string, args, locale) => {
   });
 
   return finalString;
-};
+}
